add some tests, refactor
