+++ e2e/integration/utils/utils.js
@@ -1,0 +1,15 @@
+const delay = (duration)=> new Promise(resolve=>setTimeout(resolve,duration));
+
+module.exports.pollUntil = async (action, assert, delayDuration = 0) =>{
+    while (true){
+       let result = await action();
+       try{
+          assert(result)
+          break;
+       }
+       catch (ex){
+         console.error(ex);
+       }
+       await delay(delayDuration)
+    }
+}
